backend/services/calendar_service.py
# backend/services/calendar_service.py
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build

class CalendarService:
    def create_event(self, title, start_time, end_time, attendees):
        # Create calendar event
        pass
    
    def check_availability(self, time_range):
        # Check free slots
        pass