backend/requirements.txt
# backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
langgraph==0.0.40
google-api-python-client==2.108.0
notion-client==2.2.0